Modify webpack configuration
Replace the img-loader plugin, which only minimizes images within our
project, with the imagemin-webpack-plugin, which minimizes all images
within the entire bundle, including images from included modules. Add
configuration to the UglifyJS webpack plugin that heavily optimizes and
minimizes the bundle's Javascript. Add the OrderOccurancePlugin which
organizes the contents of the webpack bundle in order of occurance
within the bundle so that the bundle loads and runs with better
performance. Additionally modify the devtool configuration so that it
uses a much easier to generate source map for dev builds and add the
devServer configuration to make it more obvious how we have configured
the development server.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>