Fix background image render and improve timeline
Fix the previously unaddressed problem caused by the arithmetic encoding
done via the imagemin plugin. The arithmetic encoding was making the
`tiled_1.jpg` file un-renderable, and the renderer as a result not only
failed to render the background, it also failed to render anything
correctly due to lacking a clear between frames without the background
draw. In addition to fixing that problem, fall-back code was added so
future problems with the same background draw won't cause additional
visual artifacts. Improve the timeline by extending it to 500 pixels and
making it's colors and drawing routines much better. It it now much
easier to read and understand the contents of the timeline bar, and it
now makes it much more clear it can be selected via usage of a new
cursor. Additionally, a minor addition to `renderer.ts` of the browser
prefixed versions of the image smoothing configuration.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>